### 2020-10-24 #2776 @whedon
[PRE REVIEW]: PRDA: An R pakcage for Prospective and Retrospective Design Analysis
**Submitting author:** @ClaudioZandonella (<a href="http://orcid.org/0000-0001-7721-6318">Claudio Zandonella Callegher</a>)
**Repository:** <a href="https://github.com/ClaudioZandonella/PRDA" target ="_blank">https://github.com/ClaudioZandonella/PRDA</a>
**Version:** v1.0.0.900
