### 2020-10-27 #2776 @cMadan
@ClaudioZandonella, before inviting reviewers I did an initial pass of the repository to see check if I noticed anything missing, and I don't see clear community guidelines.

You can see the reviewer checklist here (https://joss.readthedocs.io/en/latest/review_checklist.html), this is the exact same list that reviewers would need to evaluate your submission on. Please take a look and make sure that you think all of these are reasonably addressed in your submission.

Let me know after you've revised accordingly and I'll move forward with the review process.
