### 2020-10-27 #2776 @ClaudioZandonella
@cMadan, thanks! I just added a Contributor Code of Conduct link.

I checked all the other points. Probably the only other missing information is a clear statement about contribution and authorship. The other authors gave fundamental contributions to the package development in the theoretical aspects, design phase, and paper writing. However, I coded the most part so considering git commits it seems that I did most of the work although all the authors contributed in fundamental aspects.

I would like to not add verbose information in the read me about contribution and authorship. Instead, I could discuss it with the reviewers if required.
