### 2020-11-01 #2776 @ejolly
@cMadan Unfortunately I don't do much R work these days so I don't think I'd be a good reviewer. I actually requested an edit to the languages I'm listed for on a [different review](https://github.com/openjournals/joss-reviews/issues/2471#issuecomment-701711404) just to make sure I'm not a bottleneck for R package submissions in the future. Maybe that didn't get updated?
