### 2020-11-03 ClaudioZandonella/PRDA#2 @DominiqueMakowski
JOSS comment: CRAN
Related to https://github.com/openjournals/joss-reviews/issues/2810 :

It seems like this package is not on CRAN. Is this correct? If so, I wonder why is that, it seems in pretty good shape to be uploaded?
