### 2020-11-03 ClaudioZandonella/PRDA#3 @DominiqueMakowski
JOSS comment: Community guidelines
Related to openjournals/joss-reviews#2810 :

I didn't find any community guidelines (https://github.com/ClaudioZandonella/PRDA/community), adding these could be good to guide potential contributors (in addition to the contributing section of the README).
