### 2020-11-03 ClaudioZandonella/PRDA#4 @DominiqueMakowski
JOSS comment: website link
Related to openjournals/joss-reviews#2810 :

Minor, but could you link the package's website in the "About" section of GitHub?
