### 2020-11-03 ClaudioZandonella/PRDA#5 @DominiqueMakowski
JOSS comment: auto effect_type
Related to openjournals/joss-reviews#2810 :

Currently, if let's say `effect_type` arg is left to default ("correlation") and we change the `test_method` to that of a mean difference test, it errors. However, I wonder whether it would be simpler to just have the effect_type adjust automatically based on the input of test_method, especially in straightforward cases like correlations and t-tests?
