### 2020-11-03 ClaudioZandonella/PRDA#6 @DominiqueMakowski
JOSS comment: docstrings
Related to openjournals/joss-reviews#2810 :

In the API documentation of the `sample_n2` argument, I would add the cases where this argument applies (e.g., "This argument is used when 'test_method' is ...").
