### 2020-11-03 ClaudioZandonella/PRDA#7 @DominiqueMakowski
JOSS comment: Future plans
Related to openjournals/joss-reviews#2810 :

Maybe add a "Future plans" or a section like that (if applicable) so that one can see the direction of this package's future development, and eventually know where/how to contribute.
