### 2020-11-03 #2810 @DominiqueMakowski
PRDA is an R package aiming at facilitating the implementation of design analysis, a recent framework extending traditional power analysis.

The package is very well documented, with a thorough yet accessible introduction to the conceptual framework.

Regarding its API, it is also very accessible, well-documented with good examples and clear (helped by the low number of functions).
The paper is well-written, clear and to the point.
Overall, it's a really good and useful piece of software that will improve the accessibility of design analysis.

I only have some minor comments that I made directly on the software's repo (see above ☝️) :relaxed:
