### 2020-11-03 ClaudioZandonella/PRDA#8 @mmrabe
JOSS review: Add function call to result?
Refers to JOSS submission openjournals/joss-reviews#2810

You already have a full list of function arguments which is returned with the result (`call_arguments`). For some use cases, however, it may be helpful to see the actual function call (such as from `sys.call()`) that was used to generate the result. Maybe this would even be helpful to print out in `print.design_analysis`.
