### 2020-11-03 ClaudioZandonella/PRDA#9 @mmrabe
JOSS review: Rename `x` to `n` in `effect_size`?
Refers to JOSS submission openjournals/joss-reviews#2810

Having the possibility to pass a function as `effect_size` is a very nice feature. In line with most probability distribution function in base R, would it be possible to rename the argument in the documentation to `n` so that it says that it requires a function signature of `function(n) ...`? I know it's technically already possible to do so and you state in the documentation that this argument is the number of random variates to be sampled but it may avoid confusion.
