### 2020-11-03 ClaudioZandonella/PRDA#2 @mmrabe
There is a phrase in your manuscript stating that the package is available on CRAN with something looking like a dummy to be replaced with the actual link. So it seems like you're in the process of pushing to CRAN. I'm sure you're aware but please remember to replace the dummy or remove it from the manuscript if getting the package to CRAN is not feasible right now (which would be just fine).
