### 2020-11-03 ClaudioZandonella/PRDA#10 @mmrabe
JOSS review: Consider long compute times
Refers to JOSS submission openjournals/joss-reviews#2810

Under some circumstances such as sampling `effect_size` from a distribution and/or large number of iterations `B`, the computation may take very long. As the computation takes place in your C++ code, the session may appear frozen without any option of aborting or checking the status. Please consider one or more of the following:

- Parallelizing your code., e.g. with OpenMP. Ideally, this is just one additional line of code before the loop to parallelize and one or two lines in Makevars
- Displaying some sort of progress bar or a message every x% so that the user has some idea of how long the computation may take
- Checking for SIGINT or using `R_CheckUserInterrupt()` in your loop to allow the user to abort
