### 2020-11-03 #2810 @mmrabe
The submitted paper describes the R package PRDA, which can be used for retrospective and prospective design analysis. It has significant advantages over prior R packages, such as an option for prospective design analysis and using standardized effect sizes, which are more common than unstandardized effect sizes in some scientific disciplines.

The paper, documentation, and examples are thoughtfully crafted and very straightforward. As a whole, they could possibly make the rather complex issue of power/design analysis more graspable and easy to address for researchers without advanced skills in statistical programming.

@DominiqueMakowski has already thoroughly evaluated the paper/package with an exemplary speed 😲 and made a number of suggestions, all of which I support. I made three additional minor suggestions (see above 👆). Once these have been addressed, I can fully recommend the paper for publication in JOSS.
