### 2020-11-03 #2810 @ClaudioZandonella
@DominiqueMakowski and @mmrabe thanks for your very rapid reviews and thanks for your encouraging opinions!

I had a quick look at the issues and all your suggestions are very useful. Most of them will require small changes so at the weekend I will have the time to fix everything. Thanks!
