### 2020-11-15 598355b
Fixed #5 Auto effect_type
Let the effect type be defined automatically according to test method.

Update functions, documentation, and unit tests
