+++ paper/paper.Rmd
@@ -78,7 +78,7 @@ Suppose previous results in the literature indicate correlations in this area ar
 ```{r retro}
 library(PRDA)
 retrospective(effect_size = .25, sample_n1 = 20,
-              effect_type = "correlation", test_method = "pearson", seed = 2020)
+              test_method = "pearson", seed = 2020)
 ```
 
 In the output, we have the summary information about the hypothesized population effect, the study characteristics, and the inferential risks. We obtained a statistical power of almost 20% that is associated with a Type M error of around 2.2 and a Type S error of 0.01. That means, statistical significant results are on average an overestimation of 120% of the hypothesized population effect and there is a 1% probability of obtaining a statistically significant result in the opposite direction. To know more about function arguments and examples see the function documentation and vignette.
@@ -89,8 +89,7 @@ Alternatively, if no precise information about hypothetical effect size is avail
 
 ```{r retro_dist}
 retrospective(effect_size = function(x) rnorm(x, .25, .1), sample_n1 = 20,
-              effect_type = "correlation", test_method = "pearson",
-              tl = .1, tu = .4, B = 1e3, seed = 2020)
+              test_method = "pearson", tl = .1, tu = .4, B = 1e3, seed = 2020)
 ```
 
 Consequently this time we obtained a distribution of values for power, Type M error, and Type S error. Summary information are provided in the output.
@@ -100,8 +99,7 @@ Consequently this time we obtained a distribution of values for power, Type M er
 Given the previous results, researchers might consider planning a replication study to obtain more reliable results. The function `prospective()` can be used to compute the sample size needed to obtain a given level of power (e.g., power = 80%).
 
 ```{r pro}
-prospective(effect_size = .25, power = .8,
-            effect_type = "correlation", test_method = "pearson",
+prospective(effect_size = .25, power = .8, test_method = "pearson",
             display_message = FALSE, seed = 2020)
 ```
 
