+++ tests/testthat/test_prospective.R
@@ -15,13 +15,12 @@ test_that("inputs are correctly specified", {
 
   # Redefine function to avoid specify arguments each the times
   test_prospective <- function(effect_size = .3, power = .8, ratio_n = 1,
-                         effect_type = "correlation", test_method = "pearson",
-                         alternative = "two_sided", sig_level = .05,
-                         ratio_sd = 1, B = 10, tl = -Inf, tu = Inf,
-                         B_effect = 10, sample_range = c(2, 1000),
+                         test_method = "pearson", alternative = "two_sided",
+                         sig_level = .05, ratio_sd = 1, B = 10, tl = -Inf,
+                         tu = Inf, B_effect = 10, sample_range = c(2, 1000),
                          eval_power = "mean", tol = .01,
                          display_message = FALSE, seed = 2020){
-    prospective(effect_size, power, ratio_n, effect_type, test_method,
+    prospective(effect_size, power, ratio_n, test_method,
                 alternative, sig_level, ratio_sd, B, tl, tu, B_effect,
                 sample_range, eval_power, tol, display_message, seed)
   }
@@ -125,40 +124,40 @@ test_that("inputs are correctly specified", {
   #----    Other cases    ----
 
   # coherence effect_type and test_method
-  coherence_corr <- "If  'effect_type = correlation', argument 'test_method' has to be 'pearson'"
-  expect_error(test_prospective(effect_type = "correlation", test_method = "paired"),
-               coherence_corr)
-  coherence_cohen <- "No appropriate 'test_method' for 'effect_type = cohen_d'"
-  expect_error(test_prospective(effect_type = "cohen_d", test_method = "pearson"),
-               coherence_cohen)
+  # coherence_corr <- "If  'effect_type = correlation', argument 'test_method' has to be 'pearson'"
+  # expect_error(test_prospective(effect_type = "correlation", test_method = "paired"),
+  #              coherence_corr)
+  # coherence_cohen <- "No appropriate 'test_method' for 'effect_type = cohen_d'"
+  # expect_error(test_prospective(effect_type = "cohen_d", test_method = "pearson"),
+  #              coherence_cohen)
 
   # correlation and ratio_n
-  correlation_text <- "If 'effect_type = correlation', argument 'ratio_n' is set to NULL"
-  expect_warning(test_prospective(effect_type = "correlation", test_method = "pearson",
-                                  ratio_n = 2), correlation_text)
+  correlation_text <- "If 'test_method = pearson', argument 'ratio_n' is set to NULL"
+  expect_warning(test_prospective(test_method = "pearson", ratio_n = 2),
+                 correlation_text)
 
   # one_sample and ratio_n
   one_sample_text <- "If 'test_method = one_sample', argument 'ratio_n' must be set to NULL"
-  expect_error(test_prospective(ratio_n = 3, effect_type = "cohen_d", test_method = "one_sample"),
+  expect_error(test_prospective(ratio_n = 3, test_method = "one_sample"),
                one_sample_text)
 
   # paired and ratio_n
   paired_text <- "If 'test_method = paired', argument 'ratio_n' has to be 1"
-  expect_error(test_prospective(effect_type = "cohen_d", test_method = "paired", ratio_n = 2),
+  expect_error(test_prospective(test_method = "paired", ratio_n = 2),
                  paired_text)
-  expect_error(test_prospective(effect_type = "cohen_d", test_method = "paired", ratio_n = NULL),
+  expect_error(test_prospective(test_method = "paired", ratio_n = NULL),
                paired_text)
 
[…]
