### 2020-11-15 ClaudioZandonella/PRDA#5 @ClaudioZandonella
Dear @DominiqueMakowski thanks for pointing out this issue. Initially, we wanted to be sure that the user was aware of the effect type considered; so it was required to specify both the effect_type and test_method argument. However, we agree that this is redundant and most of the times it ends up with annoying error messages.  Now, following your suggestion, we let the effect type to be set automatically and the user is required to define only the test method.
At the moment the changes are in the develop branch. We will merge everything in the master branch after the review process.
