### 2020-11-15 9be3170
fix read me install options to develop branch
