+++ README.Rmd
@@ -52,7 +52,7 @@ You can install the development version from [GitHub](https://github.com/Claudio
 ``` r
 # install.packages("devtools")
 devtools::install_github("ClaudioZandonella/PRDA",
-                         build_vignettes = TRUE)
+                         ref = "develop", build_vignettes = TRUE)
 ```
 
 
