### 2020-11-15 9952933
Fixed #6 docstrings
Specify in the functions documentation when the arguments sample_n2 and ratio_n are required
