+++ man/prospective.Rd
@@ -24,14 +24,18 @@ prospective(
 )
 }
 \arguments{
-\item{effect_size}{a numeric value or function (see details) indicating the
+\item{effect_size}{a numeric value or function (see Details) indicating the
 hypothetical population effect size.}
 
 \item{power}{a numeric value indicating the required power level.}
 
-\item{ratio_n}{a numeric value indicating the ratio between \code{sample_n1}
-and \code{sample_n2}. This argument is required in the case of two-sample
-\emph{t}-test or Welch's \emph{t}-test.}
+\item{ratio_n}{a numeric value indicating the ratio between the sample size in
+the first group and in the second group. This argument is required when
+\code{test_method} is set to \code{"two_sample"} or \code{"welch"}. In the
+case of \code{test_method = "paired"}, set \code{ratio_n} to 1. Whereas in
+the case of \code{test_method = "one_sample"}, set \code{ratio_n} to
+\code{NULL}. This argument is ignored for \code{test_method = "pearson"}.
+See Test methods section in Details.}
 
 \item{test_method}{a character string specifying the test type, must be one of
 \code{"pearson"} (default, Pearson's correlation), \code{"two_sample"}
@@ -69,7 +73,7 @@ maximum sample size of the first group (\code{sample_n1}).}
 
 \item{eval_power}{a character string specifying the function used to summarize
 the resulting distribution of power values. Must be one of "median"
-(default) or "mean". You can specify just the initial letters. See details.}
+(default) or "mean". You can specify just the initial letters. See Details.}
 
 \item{tol}{a numeric value indicating the tolerance of required power level.}
 
