### 2020-11-15 ClaudioZandonella/PRDA#6 @ClaudioZandonella
Dear @DominiqueMakowski, we specified when arguments are required in the functions documentation. This made the functions help page much more clear. Thanks for the suggestion.
