### 2020-11-15 1135ec3
Fixed #8 add function call to result
Add the raw function call to returned arguments in the list call_arguments
