+++ R/retrospective.R
@@ -58,7 +58,7 @@
 #'    \item{design_analysis}{a character string indicating the type of design
 #'    analysis: "retrospective".}
 #'    \item{call_arguments}{a list with all the arguments passed to the
-#'    function.}
+#'    function and the raw function call.}
 #'    \item{effect_info}{a list with all the information regarding the
 #'    considered hypothetical population effect size. The list includes:
 #'    \code{effect_type} indicating the type of effect; \code{effect_function}
@@ -222,6 +222,7 @@ retrospective <- function(effect_size,
 
   # Save call
   design_analysis <- "retrospective"
+  arguments_raw <- sys.call()
   call_arguments <- as.list(match_call(default = TRUE))[-1]
 
   # eval possible errors
@@ -287,7 +288,8 @@ retrospective <- function(effect_size,
 
   #----    save results    ----
   design_fit <- structure(list(design_analysis = design_analysis,
-                               call_arguments = call_arguments,
+                               call_arguments = c(call_arguments,
+                                                  arguments_raw = arguments_raw),
                                effect_info = c(effect_type = effect_type,
                                                effect_info),
                                test_info = test_info,
