### 2020-11-15 ClaudioZandonella/PRDA#8 @ClaudioZandonella
Dear @mmrabe thanks for this suggestion. We agree that it would be very useful to see the actual function call. Therefore, we added to the list `call_arguments` another field `arguments_raw` with the actual function call. This will be really useful for debugging or other checkings.

We considered printing the actual function call in the function output as well. However, we think that there are two mains problems:
1. In most cases, the function call is just before the output (e.g., Rmarkdow documents) thus this would be redundant and would look odd or even unwanted to many users (thinking of other packages as "lme4" or "brms", they do not report the function call either).
2. The actual function call could be really raw if users do not specify arguments, relying only on position, or partial completion is used. For example `retrospective(.3, 25, NULL, "pe","l", seed = 2020)` runs but it is a rather enigmatic call. Of course, this is an extreme case. For this reason, in the list `call_arguments` we saved the complete and correct list of arguments but it would be too long to print (12 or 16 arguments).

Let me know, please, if you agree.

At the moment the changes are in the develop branch. We will merge everything in the master branch after all changes are made.
