### 2020-11-15 ClaudioZandonella/PRDA#9 @ClaudioZandonella
Dear @mmrabe, thanks for this suggestion that helps to enhance clarity. We agree that calling the argument `n` (instead of `x`) makes the function use much more intuitive for users. Therefore we updated all the documentation and examples.
Thanks!

At the moment the changes are in the develop branch. We will merge everything in the master branch after all changes are made.
