### 2020-11-16 10b7b42
Fixed #9 Rename `x` to `n` in `effect_size`
