+++ R/retrospective.R
@@ -90,10 +90,10 @@
 #'
 #'  The hypothetical population effect size (\code{effect_size}) can be set to a
 #'  single value or a function that allows sampling values from a given
-#'  distribution. The function has to be defined as \code{function(x)
-#'  my_function(x, ...)}, with only one single argument \code{x} representing
-#'  the number of sampled values (e.g., \code{function(x) rnorm(x, mean = 0, sd
-#'  = 1)}; \code{function(x) sample(c(.1,.3,.5), x, replace = TRUE)}). This
+#'  distribution. The function has to be defined as \code{function(n)
+#'  my_function(n, ...)}, with only one single argument \code{n} representing
+#'  the number of sampled values (e.g., \code{function(n) rnorm(n, mean = 0, sd
+#'  = 1)}; \code{function(n) sample(c(.1,.3,.5), n, replace = TRUE)}). This
 #'  allows users to define hypothetical effect size distribution according to
 #'  their needs.
 #'
@@ -170,9 +170,9 @@
 #' \dontrun{
 #' # Define effect_size using functions (long computational times)
 #' # Remember to adjust B
-#' retrospective(effect_size = function(x) rnorm(x, .3, .1), sample_n1 = 25,
+#' retrospective(effect_size = function(n) rnorm(n, .3, .1), sample_n1 = 25,
 #'               test_method = "pearson", tl = .15, B = 1e3, seed = 2020)
-#' retrospective(effect_size = function(x) rnorm(x, .3, .1), sample_n1 = 25,
+#' retrospective(effect_size = function(n) rnorm(n, .3, .1), sample_n1 = 25,
 #'               test_method = "one_sample", tl = .2, tu = .4, B = 1e3,
 #'               seed = 2020)
 #' }
