### 2020-11-16 ClaudioZandonella/PRDA#8 @mmrabe
Your current solution sounds great to me. 👍
