### 2020-11-21 ClaudioZandonella/PRDA#8 @ClaudioZandonella
Perfect! So I can consider the issue solved 👍
