### 2020-11-21 aec9ce7
Fixed #10 Consider long compute times - add progress bar
add pbapply package to create progress bar in the lappy loop.

Make the use of progress bar conditional to display_message argument in the simulate_analysis function.

Add zzz.R with onAttach options for pbapply progress bar.

Update realted documentation
