+++ paper/paper.Rmd
@@ -77,7 +77,7 @@ Suppose previous results in the literature indicate correlations in this area ar
 
 ```{r retro}
 library(PRDA)
-retrospective(effect_size = .25, sample_n1 = 20,
+retrospective(effect_size = .25, sample_n1 = 20, 
               test_method = "pearson", seed = 2020)
 ```
 
@@ -89,7 +89,8 @@ Alternatively, if no precise information about hypothetical effect size is avail
 
 ```{r retro_dist}
 retrospective(effect_size = function(n) rnorm(n, .25, .1), sample_n1 = 20,
-              test_method = "pearson", tl = .1, tu = .4, B = 1e3, seed = 2020)
+              test_method = "pearson", tl = .1, tu = .4, B = 1e3, 
+              seed = 2020, display_message = FALSE)
 ```
 
 Consequently this time we obtained a distribution of values for power, Type M error, and Type S error. Summary information are provided in the output.
