+++ R/PRDA-package.R
@@ -70,6 +70,7 @@
 #' @importFrom stats rnorm t.test cor.test qt pt sd var cor median
 #' @importFrom MASS mvrnorm
 #' @importFrom utils capture.output
+#' @importFrom pbapply pblapply
 #'
 #' @docType package
 #' @name PRDA
