### 2020-11-21 ClaudioZandonella/PRDA#10 @ClaudioZandonella
Dear @mmrabe thanks for underlining this issue. You are absolutely right, in the case of a large number of iterations the computation may tale very long. Without any messages, users don't know if the function is still running or stuck.

Considering your suggestions, all three are important improvements. However, I have currently implemented only a progress bar.

**Parallelizing the code**. This would be very important. I have seen that using OpenMP is quite straightforward but it also requires more dependencies and configurations that could be problematic on different OS. I am not an expert in C++ (actually this was my first time coding in C++) so, before implementing this, I would prefer to study more about parallelization to be sure I don't break the code. I will add the parallelization in the list of future plans (see issue #7) and I am sure once implemented it will be a great plus.

**Checking for SIGINT**. Probably I miss something here. I agree it is important to allow users to abort running computations. But this can be already obtained with the stop button in R- studio or "Ctrl-C" in the console. This already kills the Rsession process and looking at the task manager there are no other processes that keep running. So probably is not necessary to add a further check in the C++ code. But is likely that I miss something, so let me know if I am wrong.

**Progress bar**. I added a progress bar using the package `{pbapply}`.  For the progress bar,  I considered the `lapply()` loop on the sampled `effect_size `, as the other `for` loop is internal to the C++ code. As a consequence, the progress is displayed only when a function is used to define the `effect_size `. This is reasonable as long computational times occur in this case. For a single value of `effect_size `, instead, no progress bar is displayed (there is only one step in the progress bar so I hidden it) but computational times are usually less than 1-2 seconds for 10'000 iterations or less than 10 seconds for 100'000 iterations. Computational time increases for a greater number of iterations but using more than 10'000 iterations is unlikely as results are already stable. Finally, the display of the progress bar is conditional to the argument `display_message = TRUE` for both functions (`prospective` and `retrospective`).

Thanks again for your suggestions, as this improved a lot the usability of the package. Let me know if there are further issues on this topic o I can close it!
