### 2020-11-21 614e174
Fixes #3 Community guidelines, Fixes #7 Future plans;
Add community guidelines and code of conduct in the repository

Add Future plans in the read me and community guidelines
