+++ CONTRIBUTING.md
@@ -1,0 +1,20 @@
+# Contributing to PRDA
+
+The PRDA package is still in the early stages of its life. Thus, surely there are many bugs to fix and features to propose. Anyone is welcome to contribute to the PRDA package.
+
+Please note that this project is released under a [Contributor Code of Conduct](https://www.contributor-covenant.org/). By contributing to this project, you agree to abide by its terms.
+
+## How to report or fix a Bug
+
+In the case of unexpected behaviours or errors, open an issue. Try to describe with adequate details the problem and, if possible, provide a minimal reproducible example. This will be very useful to solve the problem.
+
+Alternatively, if you want to directly propose a solution, open a pull request. This should concern, however, only small fixes (e.s. documentation or typos). In the case of solutions involving a large amount of changes, it is better to open an issue to discuss and plan them before. 
+
+## How to propose new features
+
+The PRDA package is still in the early stages of its life, so if you want to propose a new feature first open an issue to discuss and plan them. See the Future Plan list below for possible new features.
+
+### Future Plans
+
+- Improve compute time by parallelizing the code
+- Implement design analysis in the case of linear regression models
