### 2020-11-21 ClaudioZandonella/PRDA#7 @ClaudioZandonella
Dear @DominiqueMakowski,  good suggestion! I added a Future Plans list in the Read me and in the Community Guidelines.

Currently, there are only two points but it will be useful in the future to collect other suggestions for new features as well.

Changes are all in the develop branch I will merge everything at the end. Let me know if this is fine so I will close the issue👍
