### 2020-11-21 ClaudioZandonella/PRDA#3 @ClaudioZandonella
Dear @DominiqueMakowski,

I added the Community Guidelines and Code of Conduct as standalone files.

Changes are all in the develop branch I will merge everything at the end. Let me know if this is fine so I will close the issue👍
