### 2020-11-21 ClaudioZandonella/PRDA#2 @ClaudioZandonella
Dear @DominiqueMakowski and @mmrabe, you are right the package is not yet on CRAN. I preferred to wait for the reviews so I could make the last changes and then send to CRAN.

If you agree with the solutions provided for all the issues, now I can merge everything in the master branch and send the package to CRAN. If accepted I will update the installation instructions in the documentation and in the paper.
