### 2020-11-23 ClaudioZandonella/PRDA#2 @mmrabe
You can always add references/citations to a CRAN package by uploading a new version but you won't be able to change a paper once it is published. And as far as I can tell, the package is a very suitable CRAN candidate already. So yes, I highly recommend submitting to CRAN soon so that you can include a CRAN reference in your manuscript.
