### 2020-11-23 ClaudioZandonella/PRDA#10 @mmrabe
As I said above, one of the three solutions would be sufficient. Your progress bar solution seems fine, so you can consider this issue solved with respect to the JOSS review.

Concerning the SIGINT suggestion: You are right that it is always possible to kill the R session but this may not be the most preferable option for the user. Killing the session also usually means dumping the entire workspace, which may contain other important computation results. When clicking the STOP button in R Studio or hitting Ctrl+C, the user typically expects that only the currently running function call is aborted, not the entire R session. As far as I can tell, PRDA function calls usually don't take very long and as long as control is handed back to R every now and then, the user will be able to actually abort the PRDA call. But if for some reason the computation takes very long because of many C++ loop iterations, checking for a SIGINT inside the C++ loop may be useful to let the user get out of a long computation without losing the whole R session.
