+++ README.Rmd
@@ -106,7 +106,7 @@ Please note that this project is released under a [Contributor Code of Conduct](
 
 #### Bugs and New Features
 
-To propose a new feature or to report a bug, please open an issue on [GitHub](https://github.com/ClaudioZandonella/PRDA/issues). See [Community guidelines](https://github.com/ClaudioZandonella/PRDA/CONTRIBUTING.md).
+To propose a new feature or to report a bug, please open an issue on [GitHub](https://github.com/ClaudioZandonella/PRDA/issues). See [Community guidelines](https://github.com/ClaudioZandonella/PRDA/blob/develop/CONTRIBUTING.md).
 
 #### Future Plans
 
