### 2020-11-28 a967ef3
add test on devel and release r version
