+++ appveyor.yml
@@ -19,8 +19,8 @@ environment:
     USE_RTOOLS: true
     NOT_CRAN: true
     R_VERSION:
-      - release
-      - devel
+      - 'release'
+      - 'devel'
   # env vars that may need to be set, at least temporarily, from time to time
   # see https://github.com/krlmlr/r-appveyor#readme for details
   # USE_RTOOLS: true
