### 2020-11-28 39dfc37
Add r develop and release for travis ci
