+++ .travis.yml
@@ -14,16 +14,6 @@ jobs:
           - $HOME/.local/share/renv # global renv cache in linux (not used)
           - $TRAVIS_BUILD_DIR/renv/library # local renv cache
 
-    - name: "Linux R develop" # Linux R develop
-      os: linux
-      language: R
-      r: devel
-      cache:
-        directories:
-          - /home/travis/.linuxbrew
-          - $HOME/.local/share/renv # global renv cache in linux (not used)
-          - $TRAVIS_BUILD_DIR/renv/library # local renv cache
-
     - name: "macOS R release"       # MacOS R release
       os: osx
       language: R
@@ -46,30 +36,6 @@ jobs:
       before_cache:
         - brew cleanup
         - find /usr/local/Homebrew \! -regex ".+\.git.+" -delete
-
-   - name: "macOS R develop"      # MacOS R develop
-      os: osx
-      language: R
-      r: devel
-      r_build_args: '--no-build-vignettes'
-      r_check_args: '--ignore-vignettes'
-      env:
-        - RENV_PATHS_ROOT="$HOME/renv/cache"
-      before_install:
-        - brew update
-        # - brew install ccache
-      cache:
-        ccache: true
-        directories:
-          - $HOME/Library/Caches/Homebrew
-          - /usr/local/Homebrew
-          - /usr/local/lib/R
-          - $RENV_PATHS_ROOT # global renv cache in MacOS (not used)
-          - $TRAVIS_BUILD_DIR/renv/library # local renv cache
-      before_cache:
-        - brew cleanup
-        - find /usr/local/Homebrew \! -regex ".+\.git.+" -delete
-
 
 
 install:
