+++ .travis.yml
@@ -47,6 +47,29 @@ jobs:
         - brew cleanup
         - find /usr/local/Homebrew \! -regex ".+\.git.+" -delete
 
+    - name: "macOS R devel"       # MacOS R release
+      os: osx
+      language: R
+      r: devel
+      r_build_args: '--no-build-vignettes'
+      r_check_args: '--ignore-vignettes'
+      env:
+        - RENV_PATHS_ROOT="$HOME/renv/cache"
+      before_install:
+        - brew update
+        # - brew install ccache
+      cache:
+        ccache: true
+        directories:
+          - $HOME/Library/Caches/Homebrew
+          - /usr/local/Homebrew
+          - /usr/local/lib/R
+          - $RENV_PATHS_ROOT # global renv cache in MacOS (not used)
+          - $TRAVIS_BUILD_DIR/renv/library # local renv cache
+      before_cache:
+        - brew cleanup
+        - find /usr/local/Homebrew \! -regex ".+\.git.+" -delete
+
 
 install:
   - Rscript -e "if (!requireNamespace('renv', quietly = TRUE)) install.packages('renv')"
