### 2020-11-28 b831ac5
keep only release version for appveyor
