+++ appveyor.yml
@@ -22,8 +22,6 @@ environment:
   matrix:
   - R_VERSION: release
 
-  - R_VERSION: devel
-
   # env vars that may need to be set, at least temporarily, from time to time
   # see https://github.com/krlmlr/r-appveyor#readme for details
   # USE_RTOOLS: true
