+++ .travis.yml
@@ -76,5 +76,5 @@ install:
   - Rscript -e "renv::restore()"
 
 after_success:
-  - Rscript -e 'covr::codecov(token = "549179f1-fef6-4a05-be11-e2b5029df6a0", branch = "master")'
+  - Rscript -e 'covr::codecov(token = "549179f1-fef6-4a05-be11-e2b5029df6a0", branch = "develop")'
 
