### 2020-11-28 d9a78ed
add  comments for cran and test devtools
