+++ DESCRIPTION
@@ -1,42 +1,28 @@
 Package: PRDA
 Title: Conduct a Prospective or Retrospective Design Analysis
 Version: 1.0.0.9000
-Authors@R: 
-    c(person(given = "Claudio",
-             family = "Zandonella Callegher",
-             role = c("aut", "cre"),
-             email = "claudiozandonella@gmail.com",
-             comment = c(ORCID = "0000-0001-7721-6318")),
-      person(given = "Massimiliano",
-             family = "Pastore",
-             role = "aut",
-             email = "massimiliano.pastore@unipd.it",
-             comment = c(ORCID = "0000-0002-7922-6365")),
-      person(given = "Angela",
-             family = "Andreella",
-             role = "aut",
-             email = "a.andreella8@gmail.com",
-             comment = c(ORCID = "0000-0002-1141-3041")),
-      person(given = "Anna",
-             family = "Vesely",
-             role = "aut",
-             email = "anna.vesely92@gmail.com",
-             comment = c(ORCID = "0000-0001-6696-2390")),
-      person(given = "Enrico",
-             family = "Toffalini",
-             role = "aut",
-             email = "enrico.toffalini@yahoo.it",
-             comment = c(ORCID = "0000-0002-1404-5133")),
-      person(given = "Giulia",
-             family = "Bertoldo",
-             role = "aut",
-             email = "giuliabertoldo94@gmail.com",
-             comment = c(ORCID = "0000-0002-6960-3980")),
-      person(given = "Gianmarco",
-             family = "Altoè",
-             role = "aut",
-             email = "gianmarco.altoe@unipd.it",
-             comment = c(ORCID = "0000-0003-1154-9528")))
+Authors@R: c(
+    person("Claudio", "Zandonella Callegher", role = c("aut", "cre"),
+           email = "claudiozandonella@gmail.com",
+           comment = c(ORCID = "0000-0001-7721-6318")),
+    person("Massimiliano", "Pastore", role = c("aut"),
+           email = "massimiliano.pastore@unipd.it",
+           comment = c(ORCID = "0000-0002-7922-6365")),
+    person("Angela", "Andreella", role = c("aut"),
+           email = "a.andreella8@gmail.com",
+           comment = c(ORCID = "0000-0002-1141-3041")),
+    person("Anna", "Vesely", role = c("aut"),
+           email = "anna.vesely92@gmail.com",
+           comment = c(ORCID = "0000-0001-6696-2390")),
+    person("Enrico", "Toffalini", role = c("aut"),
+           email = "enrico.toffalini@yahoo.it",
+           comment = c(ORCID = "0000-0002-1404-5133")),
+    person("Giulia", "Bertoldo", role = c("aut"),
+           email = "giuliabertoldo94@gmail.com",
+           comment = c(ORCID = "0000-0002-6960-3980")),
+    person("Gianmarco", "Altoè", role = c("aut"),
+           email = "gianmarco.altoe@unipd.it",
+           comment = c(ORCID = "0000-0003-1154-9528")))
 Description: An implementation of the "Design Analysis" proposed by 
     Gelman and Carlin (2014). It combines the evaluation of Power-Analysis 
     with other inferential-risks as Type-M error (i.e. Magnitude) and Type-S
