### 2020-11-28 a61faef
ready first release CRAN
