+++ vignettes/PRDA.bib
@@ -50,7 +50,6 @@
   date = {2020},
   publisher = {{Cambridge University Press}},
   location = {{Cambridge}},
-  doi = {10.1017/9781139161879},
   collection = {Analytical Methods for Social Research},
   file = {/Users/claudio/Zotero/storage/78GJ64NB/Gelman et al. - Regression and Other Stories.pdf},
   keywords = {To read},
