### 2020-11-28 e76a9e1
corrige verison for cran
