### 2020-11-28 ClaudioZandonella/PRDA#2 @ClaudioZandonella
Dear @mmrabe and @DominiqueMakowski I have just submitted the package to CRAN. It shouldn't take too long for the CRAN review process. As soon as they answer me I will update the last changes to the manuscript. Thanks!
