### 2020-11-29 2cfe199
update cran comments
