### 2020-12-05 6b4802c
remove seed argument update test
