+++ tests/testthat/test_sample_effect.R
@@ -18,7 +18,7 @@ error_tol <- "Truncation requires too long computational time, consider possible
 test_that("evaluate sample_effect", {
   expect_equal(with_seed(2020, sample_effect(FUN = function(x) rnorm(x),
                                              B_effect = 100)$effect_samples),
-               with_seed(2020, rnorm(100)))
+               with_seed(2020, rnorm(110)[11:110]))
 
   expect_error(sample_effect(FUN = function(x,y) rnorm(x,y), B_effect = 100))
   expect_error(sample_effect(FUN = function(x) x, B_effect = 100))
