### 2020-12-05 6cce235
Update documentation after removing seed
