+++ paper/paper.Rmd
@@ -8,7 +8,7 @@ tags:
   - Type S error
   - replicabiliyt
 authors:
-  - name: Claudio Zandonella Callegher^[Corresponding author add info]
+  - name: Claudio Zandonella Callegher^[claudiozandonella@gmail.com]
     orcid: 0000-0001-7721-6318
     affiliation: 1
   - name: Giulia Bertoldo
@@ -77,8 +77,10 @@ Suppose previous results in the literature indicate correlations in this area ar
 
 ```{r retro}
 library(PRDA)
-retrospective(effect_size = .25, sample_n1 = 20, 
-              test_method = "pearson", seed = 2020)
+
+set.seed(2020) # set seed to make results reproducible
+
+retrospective(effect_size = .25, sample_n1 = 20, test_method = "pearson")
 ```
 
 In the output, we have the summary information about the hypothesized population effect, the study characteristics, and the inferential risks. We obtained a statistical power of almost 20% that is associated with a Type M error of around 2.2 and a Type S error of 0.01. That means, statistical significant results are on average an overestimation of 120% of the hypothesized population effect and there is a 1% probability of obtaining a statistically significant result in the opposite direction. To know more about function arguments and examples see the function documentation and vignette.
@@ -90,7 +92,7 @@ Alternatively, if no precise information about hypothetical effect size is avail
 ```{r retro_dist}
 retrospective(effect_size = function(n) rnorm(n, .25, .1), sample_n1 = 20,
               test_method = "pearson", tl = .1, tu = .4, B = 1e3, 
-              seed = 2020, display_message = FALSE)
+              display_message = FALSE)
 ```
 
 Consequently this time we obtained a distribution of values for power, Type M error, and Type S error. Summary information are provided in the output.
@@ -101,10 +103,10 @@ Given the previous results, researchers might consider planning a replication st
 
 ```{r pro}
 prospective(effect_size = .25, power = .8, test_method = "pearson",
-            display_message = FALSE, seed = 2020)
+            display_message = FALSE)
 ```
 
-In the output, we have again the summary information about the hypothesized population effect, the study characteristics, and the inferential risks. To obtain a power of around 80% the required sample size is $n = 126$, the associated Type M error is around 1.10 and the Type S error is approximately 0. To know more about function arguments and examples see the function documentation and vignette.
+In the output, we have again the summary information about the hypothesized population effect, the study characteristics, and the inferential risks. To obtain a power of around 80% the required sample size is $n = 122$, the associated Type M error is around 1.10 and the Type S error is approximately 0. To know more about function arguments and examples see the function documentation and vignette.
 
 In `PRDA` there are no implemented functions to obtain graphical representations of the results. However, it is easy to access all the results and use them to create the plots according to your own needs and preferences. See vignettes for an example.
 
