### 2020-12-05 98c9050
update cran comments
