+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: PRDA
 Title: Conduct a Prospective or Retrospective Design Analysis
-Version: 1.0.0
+Version: 1.0.0.900
 Authors@R: 
     c(person(given = "Claudio",
              family = "Zandonella Callegher",
