### 2020-12-05 ClaudioZandonella/PRDA#2 @ClaudioZandonella
Dear @mmrabe  and @DominiqueMakowski, I received only one note from CRAN reviewers.  They required to avoid modify the .GlobalEnv as this is not allowed by the CRAN policies. So I removed the seed argument from the functions and added external `set.seed()` in the examples and documentation to obtain reproducible results. See commit 6b4802cb9c0316aa1c05d831e161efe5d466a1a3 and 6cce235c139f8bba2537e8a61eb9d086be645511. I have already resubmitted the package to CRAN.
