### 2020-12-07 #2810 @ClaudioZandonella
Dear @cMadan I have integrated all the useful suggestions and comments by @mmrabe and @DominiqueMakowski.

The only missing issue is ClaudioZandonella/PRDA#2 regarding the upload of the package to CRAN. I have re-submitted the package to CRAN after they required a few changes to abide by their policy. In a few days, everything should be ok and I will be able to ultimate the revision.
