### 2020-12-08 3570f7d
add CRAN badge and install info
