+++ README.Rmd
@@ -23,6 +23,7 @@ library(PRDA)
 
 <!-- badges: start -->
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![CRAN status](https://www.r-pkg.org/badges/version/PRDA)](https://CRAN.R-project.org/package=PRDA)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/ClaudioZandonella/PRDA?branch=develop&svg=true)](https://ci.appveyor.com/project/ClaudioZandonella/PRDA/branch/develop)
 [![Travis build status](https://travis-ci.org/ClaudioZandonella/PRDA.svg?branch=develop)](https://travis-ci.org/ClaudioZandonella/PRDA)
 [![Codecov test coverage](https://codecov.io/gh/ClaudioZandonella/PRDA/branch/develop/graph/badge.svg)](https://codecov.io/gh/ClaudioZandonella/PRDA/branch/develop)
@@ -39,15 +40,13 @@ All the documentation is available at https://claudiozandonella.github.io/PRDA/.
 
 ## Installation
 
-<!-- You can install the released version of PRDA from [CRAN](https://CRAN.R-project.org) with: -->
+You can install the released version of PRDA from [CRAN](https://CRAN.R-project.org/package=PRDA) with:
 
-<!-- ``` r -->
-<!-- install.packages("PRDA") -->
-<!-- ``` -->
+``` r
+install.packages("PRDA")
+```
 
-<!-- And the development version from [GitHub](https://github.com/) with: -->
-
-You can install the development version from [GitHub](https://github.com/ClaudioZandonella/PRDA/tree/master) with:
+And the development version from [GitHub](https://github.com/ClaudioZandonella/PRDA/tree/master) with:
 
 ``` r
 # install.packages("devtools")
