### 2020-12-09 ClaudioZandonella/PRDA#2 @ClaudioZandonella
Dear @DominiqueMakowski and @mmrabe, now the package is available on CRAN! I have already updated all the info in the documentation and paper about downloading the package (commit 3570f7d9c2b51e6cb8839d0b1b02258e90fab5f5). So I can consider the issue closed. Thanks!
