### 2020-12-09 #2810 @ClaudioZandonella
Dear @cMadan, I have reviewed all the issues and updated the [paper/paper.md](https://github.com/ClaudioZandonella/PRDA/blob/master/paper/paper.md) file.

Thanks, @mmrabe and @DominiqueMakowski  I really appreciate your comments and suggestions. They helped me to improve the overall quality of the package.
