### 2020-12-09 65c60bc
fix issue brew update remove
