+++ .travis.yml
@@ -32,9 +32,9 @@ jobs:
       r_check_args: '--ignore-vignettes'
       env:
         - RENV_PATHS_ROOT="$HOME/renv/cache"
-      before_install:
-        - brew update
-        # - brew install ccache
+      # before_install:
+      #   - brew update
+      #   - brew install ccache
       cache:
         ccache: true
         directories:
@@ -55,9 +55,9 @@ jobs:
       r_check_args: '--ignore-vignettes'
       env:
         - RENV_PATHS_ROOT="$HOME/renv/cache"
-      before_install:
-        - brew update
-        # - brew install ccache
+      # before_install:
+      #   - brew update
+      #   - brew install ccache
       cache:
         ccache: true
         directories:
