### 2021-02-12 #2810 @ClaudioZandonella
Dear @cMadan, thanks for your work.

Here the last required information:

- tag version v1.0.0 [https://github.com/ClaudioZandonella/PRDA/releases/tag/v1.0.0](https://github.com/ClaudioZandonella/PRDA/releases/tag/v1.0.0)
- Zenodo DOI [http://doi.org/10.5281/zenodo.4533739](http://doi.org/10.5281/zenodo.4533739)

Let me know if there is something else I can do.
Thanks!
